requirements.txt
Django==5.1.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
pillow==12.1.1
